backend/requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==1.11
django-allauth==0.38.0
django-braces==1.13.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
